requirements.txt
Django
djangorestframework
psycopg2-binary
drf-yasg
channels[daphne]
PyJWT==2.9.0
django-cors-headers